+++ test/actions/BackendAction-test.js
@@ -5,9 +5,10 @@ const fsEx = require('fs-extra')
 const proxyquire = require('proxyquire')
 
 const utils = require('../../lib/utils/utils')
-
-const UserSettings = require('../../lib/user/UserSettings')
+const { SETTINGS_FOLDER, EXTENSIONS_FOLDER } = require('../../lib/app/Constants')
 const AppSettings = require('../../lib/app/AppSettings')
+const UserSettings = require('../../lib/user/UserSettings')
+const DcHttpClient = require('../../lib/DcHttpClient')
 const userSettingsFolder = path.join('build', 'usersettings')
 const appPath = path.join('build', 'appsettings')
 
@@ -15,74 +16,72 @@ describe('BackendAction', () => {
   /**
    * @type {BackendAction}
    */
-  let backendAction
+  let subjectUnderTest
   let logger = {}
   const BackendAction = proxyquire('../../lib/actions/BackendAction', {
     '../logger': logger
   })
 
-  let userSettings
   let appSettings
+  let userSettings
+  let dcHttpClient
 
-  before(() => {
+  before(async () => {
     process.env.USER_PATH = userSettingsFolder
-    process.env.APP_PATH = appPath
-    fsEx.emptyDir(userSettingsFolder)
-      .then(() => fsEx.emptyDir(path.join(appPath, AppSettings.SETTINGS_FOLDER)))
+    await fsEx.emptyDir(userSettingsFolder)
+    await fsEx.emptyDir(path.join(appPath, SETTINGS_FOLDER))
   })
 
-  beforeEach(function () {
+  beforeEach(async () => {
     process.env.USER_PATH = userSettingsFolder
 
-    return fsEx.emptyDir(userSettingsFolder)
-      .then(() => fsEx.emptyDir(path.join(appPath, AppSettings.SETTINGS_FOLDER)))
-      .then(() => {
-        userSettings = new UserSettings().setToken({})
-        appSettings = new AppSettings().setId('foobarTest')
+    await fsEx.emptyDir(userSettingsFolder)
+    await fsEx.emptyDir(path.join(appPath, SETTINGS_FOLDER))
 
-        backendAction = new BackendAction()
+    appSettings = await new AppSettings(appPath).setId('foobarTest')
+    appSettings.loadAttachedExtensions = sinon.stub().resolves()
+    userSettings = await new UserSettings().setToken({})
+    dcHttpClient = new DcHttpClient(userSettings, null)
 
-        backendAction.userSettings = userSettings
-        backendAction.appSettings = appSettings
-      })
-      .then(() => fsEx.emptyDir(backendAction.pipelinesFolder))
-      .then(() => {
-        backendAction.pipelineWatcher = {
-          start: sinon.stub().resolves(),
-          close: () => {},
-          on: (event, fn) => fn(event, path.join(backendAction.pipelinesFolder, 'testPipeline.json'))
-        }
-        backendAction.extensionConfigWatcher = {
-          start: () => sinon.stub().resolves(),
-          on: () => sinon.stub().resolves(),
-          stop: () => sinon.stub().resolves()
-        }
+    subjectUnderTest = new BackendAction(appSettings, userSettings, dcHttpClient)
 
-        backendAction.cliProxy = {
-          start: () => sinon.stub().resolves(),
-          close: () => sinon.stub().resolves()
-        }
+    await fsEx.emptyDir(subjectUnderTest.pipelinesFolder)
 
-        backendAction.attachedExtensionsWatcher = {
-          attachedExtensions: []
-        }
+    subjectUnderTest.pipelineWatcher = {
+      start: sinon.stub().resolves(),
+      close: () => {},
+      on: (event, fn) => fn(event, path.join(subjectUnderTest.pipelinesFolder, 'testPipeline.json'))
+    }
+    subjectUnderTest.extensionConfigWatcher = {
+      start: sinon.stub().resolves(),
+      on: sinon.stub().resolves(),
+      stop: sinon.stub().resolves()
+    }
 
-        backendAction.dcClient = {}
+    subjectUnderTest.cliProxy = {
+      start: sinon.stub().resolves(),
+      close: sinon.stub().resolves()
+    }
 
-        logger.info = () => {}
-        logger.error = () => {}
-        logger.debug = () => {}
-      })
+    subjectUnderTest.attachedExtensionsWatcher = {
+      attachedExtensions: []
+    }
+
+    subjectUnderTest.dcHttpClient = {}
+
+    logger.info = () => {}
+    logger.error = () => {}
+    logger.debug = () => {}
   })
 
   afterEach(async () => {
-    backendAction.pipelineWatcher.close()
-    return backendAction.extensionConfigWatcher.stop()
+    await subjectUnderTest.pipelineWatcher.close()
+    await subjectUnderTest.extensionConfigWatcher.stop()
   })
 
-  after(() => {
-    return fsEx.remove(userSettingsFolder)
-      .then(() => fsEx.remove(appPath))
+  after(async () => {
+    await fsEx.remove(userSettingsFolder)
+    await fsEx.remove(appPath)
   })
 
   describe('general', () => {
@@ -91,30 +90,32 @@ describe('BackendAction', () => {
       caporal.command = sinon.stub().returns(caporal)
       caporal.description = sinon.stub().returns(caporal)
       caporal.action = sinon.stub().returns(caporal)
+      caporal.option = sinon.stub().returns(caporal)
 
-      BackendAction.register(caporal)
+      BackendAction.register(caporal, appSettings, userSettings)
 
       assert(caporal.command.calledWith('backend start'))
       assert(caporal.description.calledOnce)
       assert(caporal.action.calledOnce)
     })
 
-    it('should throw if user not logged in', (done) => {
-      userSettings.setToken()
-      backendAction.run()
-        .catch(err => {
-          assert.equal(err.message, 'You\'re not logged in! Please run `sgcloud login` again.')
-          done()
-        })
+    it('should throw if user not logged in', async () => {
+      await userSettings.setToken()
+      try {
+        await subjectUnderTest.run()
+        assert.fail('Expected error to be thrown.')
+      } catch (err) {
+        assert.equal(err.message, 'You\'re not logged in! Please run `sgcloud login` again.')
+      }
     })
 
     it('should fail because a backend process is already running', (done) => {
       const pid = process.pid
-      const processFile = path.join(appPath, AppSettings.SETTINGS_FOLDER)
+      const processFile = path.join(appPath, SETTINGS_FOLDER)
 
       utils.setProcessFile('backend', processFile, pid)
 
-      backendAction.run()
+      subjectUnderTest.run()
         .catch(err => {
           assert.equal(err.message, `Backend process is already running with pid: ${pid}. Please quit this process first.`)
           done()
@@ -124,22 +125,22 @@ describe('BackendAction', () => {
 
   describe('_pipelineEvent', () => {
     it('should not do anything when the extension is not attached', (done) => {
-      backendAction._pipelineRemoved = sinon.stub().resolves()
-      backendAction._pipelineChanged = sinon.stub().resolves()
+      subjectUnderTest._pipelineRemoved = sinon.stub().resolves()
+      subjectUnderTest._pipelineChanged = sinon.stub().resolves()
 
       logger.debug = (msg) => {
         assert.equal(msg, 'The extension of the pipeline is not attached --> skip')
-        assert.ok(!backendAction._pipelineRemoved.called, '_pipelineRemoved should not be called')
-        assert.ok(!backendAction._pipelineChanged.called, '_pipelineChanged should not be called')
+        assert.ok(!subjectUnderTest._pipelineRemoved.called, '_pipelineRemoved should not be called')
+        assert.ok(!subjectUnderTest._pipelineChanged.called, '_pipelineChanged should not be called')
         done()
       }
-      backendAction._pipelineEvent('test', 'somefile')
+      subjectUnderTest._pipelineEvent('test', 'somefile')
     })
   })
 
   describe('watching', () => {
     it('should update pipelines', (done) => {
-      backendAction.backendProcess = {
+      subjectUnderTest.backendProcess = {
         connect: sinon.stub().resolves(),
         selectApplication: sinon.stub().resolves(),
         resetPipelines: sinon.stub().resolves(),
@@ -147,22 +148,22 @@ describe('BackendAction', () => {
         reloadPipelineController: sinon.stub().resolves()
       }
 
-      backendAction.dcClient = {
+      subjectUnderTest.dcHttpClient = {
         downloadPipelines: sinon.stub().resolves([{pipeline: {id: 'testPipeline'}}]),
         removePipeline: sinon.stub().resolves(),
         uploadMultiplePipelines: sinon.stub().resolves()
       }
 
-      backendAction.attachedExtensionsWatcher = {
+      subjectUnderTest.attachedExtensionsWatcher = {
         attachedExtensions: [],
         start: () => sinon.stub().resolves(),
         on: () => sinon.stub().resolves()
       }
 
-      backendAction._extensionChanged = sinon.stub().resolves()
+      subjectUnderTest._updateExtensionConfig = sinon.stub().resolves()
 
-      backendAction._startSubProcess()
-        .then(() => fsEx.readJson(path.join(process.env.APP_PATH, 'pipelines', 'testPipeline.json')))
+      subjectUnderTest._startSubProcess()
+        .then(() => fsEx.readJson(path.join(appPath, 'pipelines', 'testPipeline.json')))
         .then((content) => {
           assert.deepEqual(content, {pipeline: {id: 'testPipeline'}})
           done()
@@ -172,7 +173,7 @@ describe('BackendAction', () => {
     it('should fail when pipeline IDs not matching pipeline file names', (done) => {
       appSettings.loadAttachedExtensions = () => { return {testExtension: {path: '..'}} }
 
-      backendAction.backendProcess = {
+      subjectUnderTest.backendProcess = {
         connect: sinon.stub().resolves(),
         selectApplication: sinon.stub().resolves(),
         resetPipelines: sinon.stub().resolves(),
@@ -180,25 +181,25 @@ describe('BackendAction', () => {
         reloadPipelineController: sinon.stub().resolves()
       }
 
-      backendAction.dcClient = {
+      subjectUnderTest.dcHttpClient = {
         downloadPipelines: sinon.stub().resolves([]),
         removePipeline: sinon.stub().resolves(),
         uploadMultiplePipelines: sinon.stub().resolves()
       }
 
-      backendAction.attachedExtensionsWatcher = {
+      subjectUnderTest.attachedExtensionsWatcher = {
         attachedExtensions: [{'testPipeline123': {path: ''}}],
         start: () => sinon.stub().resolves(),
         on: () => sinon.stub().resolves()
       }
 
-      backendAction._extensionChanged = sinon.stub().resolves()
+      subjectUnderTest._updateExtensionConfig = sinon.stub().resolves()
 
-      fsEx.writeJson(path.join(process.env.APP_PATH, 'pipelines', 'testPipeline.json'), {pipeline: {id: 'testPipeline123'}}, err => {
+      fsEx.writeJson(path.join(appPath, 'pipelines', 'testPipeline.json'), {pipeline: {id: 'testPipeline123'}}, err => {
         assert.ifError(err)
 
-        backendAction._startSubProcess()
-          .then(() => fsEx.readJson(path.join(process.env.APP_PATH, 'pipelines', 'testPipeline.json')))
+        subjectUnderTest._startSubProcess()
+          .then(() => fsEx.readJson(path.join(appPath, 'pipelines', 'testPipeline.json')))
           .then((content) => {
             assert.deepEqual(content, {pipeline: {id: 'testPipeline123'}})
           })
@@ -214,8 +215,8 @@ describe('BackendAction', () => {
       })
     })
 
-    it('should call dcClient if pipelines were updated', (done) => {
-      backendAction.backendProcess = {
+    it('should call dcClient if pipelines were updated', async () => {
+      subjectUnderTest.backendProcess = {
         connect: sinon.stub().resolves(),
         selectApplication: sinon.stub().resolves(),
         resetPipelines: sinon.stub().resolves(),
@@ -225,38 +226,34 @@ describe('BackendAction', () => {
 
       const pipeline = {pipeline: {id: 'plFooBarline1'}}
       const appId = 'foobarAppIdDcTestBackendAction'
-      appSettings.setId(appId)
+      await appSettings.setId(appId)
 
-      const file = path.join(backendAction.pipelinesFolder, 'plFooBarline1.json')
-      assert.equal(backendAction.pipelines[file], undefined)
+      const file = path.join(subjectUnderTest.pipelinesFolder, 'plFooBarline1.json')
+      assert.equal(subjectUnderTest.pipelines[file], undefined)
 
-      backendAction.dcClient.uploadPipeline = (f, aId, trusted) => {
-        assert.deepEqual(backendAction.pipelines[file].id, pipeline.pipeline.id)
+      subjectUnderTest.dcHttpClient.uploadPipeline = (f, aId) => {
+        assert.deepEqual(subjectUnderTest.pipelines[file].id, pipeline.pipeline.id)
         assert.deepEqual(f, pipeline)
         assert.equal(aId, appId)
       }
 
-      fsEx.writeJson(file, pipeline, (err) => {
-        assert.ifError(err)
-        backendAction._pipelineChanged(file).then(() => {
-          done()
-        })
-      })
+      await fsEx.writeJson(file, pipeline)
+      await subjectUnderTest._pipelineChanged(file)
     })
 
     it('should write generated extension-config if backend-extension was updated', async () => {
       let generated = {backend: {id: 'myGeneratedExtension'}}
 
       let called = 0
-      backendAction.dcClient.generateExtensionConfig = () => {
+      subjectUnderTest.dcHttpClient.generateExtensionConfig = () => {
         called++
         return Promise.resolve(generated)
       }
 
-      const cfgPath = path.join(process.env.APP_PATH, 'extensions', 'testExt')
+      const cfgPath = path.join(appPath, 'extensions', 'testExt')
 
       try {
-        await backendAction._extensionChanged({file: generated, path: cfgPath})
+        await subjectUnderTest._updateExtensionConfig({file: generated, path: cfgPath})
         const content = await fsEx.readJson(path.join(cfgPath, 'extension', 'config.json'))
         assert.deepEqual(content, {id: 'myGeneratedExtension'})
         assert.equal(called, 1)
@@ -269,15 +266,15 @@ describe('BackendAction', () => {
       let generated = {frontend: {id: 'myGeneratedExtension'}}
 
       let called = 0
-      backendAction.dcClient.generateExtensionConfig = () => {
+      subjectUnderTest.dcHttpClient.generateExtensionConfig = () => {
         called++
         return Promise.resolve(generated)
       }
 
-      const cfgPath = path.join(process.env.APP_PATH, 'extension', 'testExt')
+      const cfgPath = path.join(appPath, 'extension', 'testExt')
 
       try {
-        await backendAction._extensionChanged({file: generated, path: cfgPath})
+        await subjectUnderTest._updateExtensionConfig({file: generated, path: cfgPath})
         const content = await fsEx.readJson(path.join(cfgPath, 'frontend', 'config.json'))
         assert.deepEqual(content, {id: 'myGeneratedExtension'})
         assert.equal(called, 1)
@@ -288,14 +285,14 @@ describe('BackendAction', () => {
 
     it('should throw error if dcClient is not reachable', (done) => {
       const pipeline = {pipeline: {id: 'dCPlTest2'}}
-      backendAction.dcClient = {
+      subjectUnderTest.dcHttpClient = {
         uploadPipeline: sinon.stub().rejects(new Error('error'))
       }
 
-      const file = path.join(backendAction.pipelinesFolder, 'dCPlTest2.json')
+      const file = path.join(subjectUnderTest.pipelinesFolder, 'dCPlTest2.json')
       fsEx.writeJson(file, pipeline, (err) => {
         assert.ifError(err)
-        backendAction._pipelineChanged(file).catch(err => {
+        subjectUnderTest._pipelineChanged(file).catch(err => {
           assert.ok(err)
           assert.equal(err.message, `error`)
           done()
@@ -305,13 +302,13 @@ describe('BackendAction', () => {
 
     it('should return if pipeline was changed', (done) => {
       const pipeline = {pipeline: {id: 'dCPlTest3'}}
-      const file = path.join(backendAction.pipelinesFolder, 'dCPlTest3.json')
+      const file = path.join(subjectUnderTest.pipelinesFolder, 'dCPlTest3.json')
 
-      backendAction.dcClient = {
+      subjectUnderTest.dcHttpClient = {
         uploadPipeline: sinon.stub().resolves()
       }
 
-      backendAction.backendProcess = {
+      subjectUnderTest.backendProcess = {
         connect: sinon.stub().resolves(),
         selectApplication: sinon.stub().resolves(),
         resetPipelines: sinon.stub().resolves(),
@@ -321,7 +318,7 @@ describe('BackendAction', () => {
 
       fsEx.writeJson(file, pipeline, (err) => {
         assert.ifError(err)
-        backendAction._pipelineChanged(file).then(err => {
+        subjectUnderTest._pipelineChanged(file).then(err => {
           assert.ifError(err)
           done()
         })
@@ -331,10 +328,10 @@ describe('BackendAction', () => {
     it('should throw an error if pipeline id is not matching with the filename', (done) => {
       const pipeline = {pipeline: {id: 'nonMatchingId'}}
 
-      const file = path.join(backendAction.pipelinesFolder, 'dCPlTest3.json')
+      const file = path.join(subjectUnderTest.pipelinesFolder, 'dCPlTest3.json')
       fsEx.writeJson(file, pipeline, (err) => {
         assert.ifError(err)
-        backendAction._pipelineChanged(file).catch(err => {
+        subjectUnderTest._pipelineChanged(file).catch(err => {
           assert.ok(err)
           assert.equal(err.message, `Pipeline ID "${pipeline.pipeline.id}" should match its file name!`)
           done()
@@ -344,11 +341,11 @@ describe('BackendAction', () => {
 
     it('should throw an error if pipeline is invalid', (done) => {
       const pipeline = {}
-      const file = path.join(backendAction.pipelinesFolder, 'dCPlTest3.json')
+      const file = path.join(subjectUnderTest.pipelinesFolder, 'dCPlTest3.json')
 
       fsEx.writeJson(file, pipeline, (err) => {
         assert.ifError(err)
-        backendAction._pipelineChanged(file).catch(err => {
+        subjectUnderTest._pipelineChanged(file).catch(err => {
           assert.ok(err)
           assert.equal(err.message, `invalid pipeline; check the pipeline.id property in ${file}`)
           done()
@@ -358,11 +355,11 @@ describe('BackendAction', () => {
 
     it('should throw an error if pipeline has invalid json', (done) => {
       const pipeline = '{someInvalid: content}'
-      const file = path.join(backendAction.pipelinesFolder, 'dCPlTest5.json')
+      const file = path.join(subjectUnderTest.pipelinesFolder, 'dCPlTest5.json')
 
       fsEx.writeFile(file, pipeline, (err) => {
         assert.ifError(err)
-        backendAction._pipelineChanged(file).catch(err => {
+        subjectUnderTest._pipelineChanged(file).catch(err => {
           assert.ok(err)
           assert.equal(err.message, 'Parse error on line 1:\n{someInvalid: content\n-^\nExpecting \'STRING\', \'}\', got \'undefined\'')
           done()
@@ -373,31 +370,75 @@ describe('BackendAction', () => {
     it('should return if pipeline was removed', (done) => {
       const pipelineId = 'dCPlTest4'
       let called = false
-      backendAction._writeLocalPipelines = sinon.mock().resolves()
-      backendAction.dcClient.removePipeline = (plId, id, trusted) => {
+      subjectUnderTest._writeLocalPipelines = sinon.mock().resolves()
+      subjectUnderTest.dcHttpClient.removePipeline = (plId, id, trusted) => {
         assert.equal(plId, pipelineId)
         called = true
       }
 
-      const file = path.join(backendAction.pipelinesFolder, 'dCPlTest4.json')
-      backendAction.pipelines[file] = {id: pipelineId}
+      const file = path.join(subjectUnderTest.pipelinesFolder, 'dCPlTest4.json')
+      subjectUnderTest.pipelines[file] = {id: pipelineId}
 
-      backendAction._pipelineRemoved(file).then(() => {
+      subjectUnderTest._pipelineRemoved(file).then(() => {
         assert.ok(called)
-        assert.ok(backendAction._writeLocalPipelines.called)
+        assert.ok(subjectUnderTest._writeLocalPipelines.called)
         done()
       }).catch(err => {
         assert.ifError(err)
       })
     })
 
-    it('should work', () => {
-      backendAction._startSubProcess = () => {}
+    it('should work', async () => {
+      appSettings.loadAttachedExtensions = () => { return {testExtension: {path: '..'}} }
+      subjectUnderTest._startSubProcess = () => {}
       try {
-        backendAction.run()
+        subjectUnderTest.writeExtensionConfigs = () => {}
+        await subjectUnderTest.run({})
       } catch (err) {
         assert.ifError(err)
       }
+    })
+
+    it('should write the extension configs for the app', async () => {
+      subjectUnderTest.appSettings.EXTENSIONS_FOLDER = 'extensions'
+      const extensionConfigPath = path.join(
+        subjectUnderTest.appSettings.getApplicationFolder(), EXTENSIONS_FOLDER,
+        'test-extension',
+        'extension-config.json'
+      )
+
+      fsEx.createFileSync(extensionConfigPath)
+      let mockConf = {someKey: 'someVal'}
+      await fsEx.writeJson(extensionConfigPath, mockConf)
+      let called = 0
+      subjectUnderTest.appSettings.getId = () => 1
+      subjectUnderTest.dcHttpClient.generateExtensionConfig = (config) => {
+        called++
+        return Promise.resolve(config)
+      }
+
+      appSettings.loadAttachedExtensions = () => { return {testExtension: {path: 'test-extension'}} }
+      subjectUnderTest._startSubProcess = () => {}
+
+      await subjectUnderTest.run({})
+      assert.ok(called !== 0, 'dcClient generateExtensionConfig should have been called at least once')
+    })
+
+    it('should pass true to StepExecutor\'s "inspect" constructor argument when called with --inspect', async () => {
+      subjectUnderTest.writeExtensionConfigs = () => {}
+      subjectUnderTest._startSubProcess = async function () {
+        if (this.backendProcess.executor.inspect !== true) throw new Error('Expected third constructor argument "inspect" to true.')
+      }
+
+      await subjectUnderTest.run({inspect: true})
+    })
+
+    it('should pass false to StepExecutor\'s "inspect" constructor argument when called without --inspect', async () => {
+      subjectUnderTest._startSubProcess = async function () {
+        if (this.backendProcess.executor.inspect !== false) throw new Error('Expected third constructor argument "inspect" to false.')
+      }
+
+      await subjectUnderTest.run({})
     })
   })
 })