+++ test/actions/FrontendAction-test.js
@@ -2,22 +2,29 @@ const assert = require('assert')
 const proxyquire = require('proxyquire')
 const path = require('path')
 const fsEx = require('fs-extra')
-const UserSettings = require('../../lib/user/UserSettings')
+const sinon = require('sinon')
 const AppSettings = require('../../lib/app/AppSettings')
+const UserSettings = require('../../lib/user/UserSettings')
+const DcHttpClient = require('../../lib/DcHttpClient')
+const { SETTINGS_FOLDER } = require('../../lib/app/Constants')
+const utils = require('../../lib/utils/utils')
 
 const appPath = path.join('build', 'appsettings')
 const userPath = path.join('build', 'usersettings')
 
 const fsExtraMock = {}
 const inquirer = {}
-const dcClientMock = class {}
 
-class FrontendActionMock {
+class FrontendProcessMock {
   run () { return Promise.resolve() }
 }
 
 describe('FrontendAction', () => {
-  let frontendAction
+  let subjectUnderTest
+  let appSettings
+  let userSettings
+  let dcHttpClient
+
   const FrontendAction = proxyquire('../../lib/actions/FrontendAction', {
     '../logger': {
       info: () => {},
@@ -25,28 +32,32 @@ describe('FrontendAction', () => {
       debug: () => {},
       warn: () => {}
     },
-    '../app/frontend/FrontendProcess': FrontendActionMock,
+    '../app/frontend/FrontendProcess': FrontendProcessMock,
     'fs-extra': fsExtraMock,
     'inquirer': inquirer,
-    '../DcHttpClient': dcClientMock
+    '../utils/utils': utils
   })
 
-  beforeEach((done) => {
+  beforeEach(async () => {
     process.env.USER_PATH = userPath
-    fsEx.emptyDirSync(userPath)
-    new UserSettings().setToken({})
-    process.env.APP_PATH = appPath
-    fsEx.emptyDirSync(path.join(appPath, AppSettings.SETTINGS_FOLDER))
+    await fsEx.emptyDir(userPath)
+    userSettings = await new UserSettings().setToken({})
+    await fsEx.emptyDir(path.join(appPath, SETTINGS_FOLDER))
+    dcHttpClient = new DcHttpClient(userSettings, null)
 
     fsExtraMock.existsSync = () => true
     fsExtraMock.readJSONSync = () => {}
     fsExtraMock.readdir = () => Promise.resolve()
     fsExtraMock.lstat = () => Promise.resolve()
 
-    new AppSettings().setId('foobarTest')
-    frontendAction = new FrontendAction()
+    appSettings = await new AppSettings(appPath).setId('foobarTest')
+    subjectUnderTest = new FrontendAction(appSettings, userSettings, dcHttpClient)
 
-    done()
+    subjectUnderTest.extensionConfigWatcher = {
+      start: () => sinon.stub().resolves(),
+      on: () => sinon.stub().resolves(),
+      stop: () => sinon.stub().resolves()
+    }
   })
 
   afterEach(() => {
@@ -54,28 +65,30 @@ describe('FrontendAction', () => {
     delete process.env.USER_PATH
   })
 
-  describe('constructor', () => {
-    it('should throw an error if user is not logged in', () => {
-      new UserSettings().setToken(null)
+  describe('constructor', async () => {
+    it('should throw an error if user is not logged in', async () => {
+      userSettings.setToken(null)
       try {
-        frontendAction = new FrontendAction()
-        assert.fail()
+        // re-create the subject under test after token was set to null
+        subjectUnderTest = new FrontendAction(appSettings, userSettings)
       } catch (err) {
+        assert.ok(err)
         assert.equal(err.message, 'You\'re not logged in! Please run `sgcloud login` again.')
       }
     })
   })
 
-  describe('run() -> start', () => {
+  describe('run() -> start', async () => {
+    await new UserSettings().setToken({})
     it('should throw an error if the theme is not existing', async () => {
       const options = {theme: 'theme-gmd'}
       fsExtraMock.readdir = (source) => Promise.resolve(['a', 'b'])
       fsExtraMock.lstat = () => Promise.resolve({isDirectory: () => true})
       fsExtraMock.exists = () => Promise.resolve(false)
-      frontendAction.dcClient.generateExtensionConfig = (configFile, id, cb) => { cb(null, {}) }
+      subjectUnderTest.dcHttpClient.generateExtensionConfig = (configFile, id, cb) => { cb(null, {}) }
 
       try {
-        await frontendAction.run('start', null, options)
+        await subjectUnderTest.run('start', null, options)
         assert.fail()
       } catch (err) {
         assert.equal(err.message, 'Can\'t find theme \'theme-gmd\'. Please make sure you passed the right theme.')
@@ -88,10 +101,14 @@ describe('FrontendAction', () => {
       fsExtraMock.readdir = () => Promise.resolve(['theme-gmd'])
       fsExtraMock.exists = () => Promise.resolve(true)
       fsExtraMock.lstat = () => Promise.resolve({isDirectory: () => true})
-      frontendAction.dcClient.generateExtensionConfig = () => Promise.resolve({})
+      subjectUnderTest.dcHttpClient.generateExtensionConfig = () => Promise.resolve({})
+
+      let gotCalled = false
+      utils.generateComponentsJson = () => { gotCalled = true }
 
       try {
-        await frontendAction.run('start', {}, options)
+        await subjectUnderTest.run('start', {}, options)
+        assert.ok(gotCalled)
       } catch (err) {
         assert.ifError(err)
       }
@@ -100,11 +117,11 @@ describe('FrontendAction', () => {
 
   describe('run() -> setup', () => {
     it('should run frontend setup', (done) => {
-      frontendAction.frontendSetup.run = () => {
+      subjectUnderTest.frontendSetup.run = () => {
         done()
         return Promise.resolve()
       }
-      frontendAction.run('setup', {theme: 'theme-gmd'})
+      subjectUnderTest.run('setup', {theme: 'theme-gmd'})
     })
   })
 })