+++ test/app/backend/extensionRuntime/StepExecutor-test.js
@@ -1,15 +1,14 @@
 const assert = require('assert')
 const path = require('path')
 const fsEx = require('fs-extra')
-const async = require('neo-async')
 const glob = require('glob')
 const sinon = require('sinon')
 const StepExecutor = require('../../../../lib/app/backend/extensionRuntime/StepExecutor')
 const AppSettings = require('../../../../lib/app/AppSettings')
 const UserSettings = require('../../../../lib/user/UserSettings')
-const utils = require('../../../../lib/utils/utils')
 const appPath = path.join('build', 'appsettings')
 const proxyquire = require('proxyquire').noPreserveCache()
+const { EXTENSIONS_FOLDER } = require('../../../../lib/app/Constants')
 
 describe('StepExecutor', () => {
   describe('watcher', () => {
@@ -25,8 +24,8 @@ describe('StepExecutor', () => {
       }
 
       const pathes = [
-        path.join(utils.getApplicationFolder(), 'extensions', '**', 'extension', '*.js'),
-        path.join(utils.getApplicationFolder(), 'extensions', '**', 'extension', '**', '*.js')
+        path.join(appPath, 'extensions', '**', 'extension', '*.js'),
+        path.join(appPath, 'extensions', '**', 'extension', '**', '*.js')
       ]
 
       const StepExecutorMocked = proxyquire('../../../../lib/app/backend/extensionRuntime/StepExecutor', {
@@ -37,7 +36,7 @@ describe('StepExecutor', () => {
           }
         }
       })
-      const stepExecutor = new StepExecutorMocked({info: () => {}}, {getApplicationFolder: utils.getApplicationFolder})
+      const stepExecutor = new StepExecutorMocked({info: () => {}}, {getApplicationFolder: () => appPath})
       stepExecutor.start = sinon.stub().resolves()
       stepExecutor.stop = () => {
         return new Promise((resolve, reject) => {
@@ -72,8 +71,8 @@ describe('StepExecutor', () => {
       }
 
       const pathes = [
-        path.join(utils.getApplicationFolder(), 'extensions', '**', 'extension', '*.js'),
-        path.join(utils.getApplicationFolder(), 'extensions', '**', 'extension', '**', '*.js')
+        path.join(appPath, 'extensions', '**', 'extension', '*.js'),
+        path.join(appPath, 'extensions', '**', 'extension', '**', '*.js')
       ]
 
       const StepExecutorMocked = proxyquire('../../../../lib/app/backend/extensionRuntime/StepExecutor', {
@@ -85,7 +84,7 @@ describe('StepExecutor', () => {
         }
       })
 
-      const stepExecutor = new StepExecutorMocked({info: () => {}}, {getApplicationFolder: utils.getApplicationFolder})
+      const stepExecutor = new StepExecutorMocked({info: () => {}}, {getApplicationFolder: () => appPath})
 
       stepExecutor.startWatcher()
       watcher.emit('ready')
@@ -100,7 +99,7 @@ describe('StepExecutor', () => {
     let appTestFolder
     let userTestFolder
 
-    beforeEach(function (done) {
+    beforeEach(async function () {
       this.timeout(10000)
       userTestFolder = path.join('build', 'usersettings')
       process.env.USER_PATH = userTestFolder
@@ -113,40 +112,44 @@ describe('StepExecutor', () => {
       executor.stepTimeout = 1000
       executor.stepLogger = {info: () => {}, error: () => {}, debug: () => {}, warn: () => {}}
 
-      new AppSettings().setId('shop_10006')._saveExtensions({'@foo/bar': {path: 'foobar'}})
+      const appSettings = await new AppSettings(appTestFolder).setId('shop_10006')
+      await fsEx.writeJson(appSettings.attachedExtensionsFile, {attachedExtensions: {'@foo/bar': {path: 'foobar'}}})
+
       new UserSettings().setToken({})
 
-      const extensionDir = path.join(appPath, AppSettings.EXTENSIONS_FOLDER, 'foobar', 'extension')
-      fsEx.emptyDirSync(extensionDir)
+      const extensionDir = path.join(appPath, EXTENSIONS_FOLDER, 'foobar', 'extension')
+      await fsEx.emptyDir(extensionDir)
 
-      glob(path.join(__dirname, 'fakeSteps', '*.js'), {}, (err, files) => {
-        assert.ifError(err)
-        async.each(files, (file, eCb) => fsEx.copy(file, path.join(extensionDir, path.basename(file)), eCb), (err) => {
-          assert.ifError(err)
-          executor.start().then(done)
-        })
-      })
+      const files = glob(path.join(__dirname, 'fakeSteps', '*.js'), {sync: true})
+      const fileCopyResults = await Promise.all(files.map(file => fsEx.copy(file, path.join(extensionDir, path.basename(file)))))
+      if (fileCopyResults) {
+        await executor.start()
+      }
     })
 
-    afterEach((done) => {
+    afterEach(async () => {
       delete process.env.SGCLOUD_DC_WS_ADDRESS
       delete process.env.APP_PATH
       delete process.env.USER_PATH
 
-      async.parallel([
-        (cb) => fsEx.remove(appTestFolder, cb),
-        (cb) => fsEx.remove(userTestFolder, cb)
-      ], err => {
+      try {
+        await Promise.all([
+          fsEx.remove(appTestFolder),
+          fsEx.remove(userTestFolder)
+        ])
+      } catch (err) {
         assert.ifError(err)
-        executor.stop().then(done)
-      })
+      } finally {
+        await executor.stop()
+      }
     })
 
-    it('should not start another childProcess when one is already running', () => {
-      executor.onExit = () => {}
-      return executor.start().catch(err => {
+    it('should not start another childProcess when one is already running', async () => {
+      try {
+        await executor.start()
+      } catch (err) {
         assert.equal(err.message, 'childProcess already running')
-      })
+      }
     })
 
     it('should call a local step action', (done) => {
@@ -154,6 +157,21 @@ describe('StepExecutor', () => {
       const stepMeta = {
         id: '@foo/bar',
         path: '@foo/bar/simple.js',
+        meta: {appId: 'shop_123'}
+      }
+      executor.onExit = () => {}
+      executor.execute(input, stepMeta, (err, output) => {
+        assert.ifError(err)
+        assert.deepEqual(output, input)
+        done()
+      })
+    })
+
+    it('should call a local promise step action', (done) => {
+      const input = {foo: 'bar'}
+      const stepMeta = {
+        id: '@foo/bar',
+        path: '@foo/bar/promise-resolve.js',
         meta: {appId: 'shop_123'}
       }
       executor.onExit = () => {}
@@ -174,6 +192,21 @@ describe('StepExecutor', () => {
       executor.onExit = () => {}
       executor.execute(input, stepMeta, (err, output) => {
         assert.deepEqual(err, Object.assign({name: 'Error', message: 'crashed ' + stepMeta.meta.appId}, input))
+        assert.equal(output, undefined)
+        done()
+      })
+    })
+
+    it('should callback error of promise step (with all fields)', (done) => {
+      const input = {foo: 'bar', bar: {nestedFoo: 'nestedBar'}}
+      const stepMeta = {
+        id: '@foo/bar',
+        path: '@foo/bar/promise-reject.js',
+        meta: {appId: 'shop_123'}
+      }
+      executor.onExit = () => {}
+      executor.execute(input, stepMeta, (err, output) => {
+        assert.equal(err.message, 'error')
         assert.equal(output, undefined)
         done()
       })
@@ -331,6 +364,32 @@ describe('StepExecutor', () => {
         assert.equal(err.code, 'ETIMEOUT')
         done()
       })
+    })
+
+    it('should start the sub process with "--inspect" if requested', async () => {
+      // stop the executor from beforeEach as we're going to instantiate our own
+      await executor.stop()
+
+      executor = new StepExecutor(log, {}, true)
+      await executor.start()
+
+      if (!executor.childProcess.spawnargs.reduce((val, current) => val || current === '--inspect', false)) {
+        await executor.stop()
+        assert.fail('Expected child process to be spawned with the "--inspect" argument.')
+      }
+    })
+
+    it('should start the sub process without "--inspect" if not requested', async () => {
+      // stop the executor from beforeEach as we're going to instantiate our own
+      await executor.stop()
+
+      executor = new StepExecutor(log, {}, false)
+      await executor.start()
+
+      if (executor.childProcess.spawnargs.reduce((val, current) => val || current === '--inspect', false)) {
+        await executor.stop()
+        assert.fail('Expected child process to be spawned with the "--inspect" argument.')
+      }
     })
   })
 })