+++ test/app/frontend/helper.js
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-const { join } = require('path')
+const {join} = require('path')
 const UserSettings = require('../../../lib/user/UserSettings')
 const AppSettings = require('../../../lib/app/AppSettings')
 const fsEx = require('fs-extra')
@@ -13,35 +13,31 @@ const fsEx = require('fs-extra')
 const userSettingsPath = join('build', 'usersettings')
 const appSettingsPath = join('build', 'appsettings')
 const appId = 'foobarTest'
+const { SETTINGS_FOLDER } = require('../../../lib/app/Constants')
 
 /**
  * Creates an application environment for tests.
- * @param {Function} callback Anything local that should be called in beforeEach afterwards.
  */
-const setupAppEnvironment = (callback = () => {}) => {
+const setupAppEnvironment = async () => {
   process.env.USER_PATH = userSettingsPath
   process.env.APP_PATH = appSettingsPath
   process.env.appId = appId
 
-  fsEx.emptyDirSync(join(appSettingsPath, AppSettings.SETTINGS_FOLDER))
-  new AppSettings().setId(appId)
+  await fsEx.emptyDir(join(appSettingsPath, SETTINGS_FOLDER))
+  await new AppSettings(appSettingsPath).setId(appId)
   new UserSettings().setToken({})
-
-  callback()
 }
 
 /**
  * Cleares the application environment after each teast.
- * @param {Function} callback Anything local that should be called in afterEach afterwards.
  */
-const clearAppEnviroment = (callback = () => {}) => {
+const clearAppEnviroment = async () => {
   delete process.env.USER_PATH
   delete process.env.APP_PATH
   delete process.env.appId
 
-  fsEx.remove(userSettingsPath, () => {
-    fsEx.remove(appSettingsPath, callback)
-  })
+  await fsEx.remove(userSettingsPath)
+  await fsEx.remove(appSettingsPath)
 }
 
 module.exports = {